Removed redundant ServiceHistoryBuilder class. Put logic in transformer
